refactor (structure) 구조변경
