+++ src/main/java/qna/domain/answer/AnswerRepository.java
@@ -1,7 +1,8 @@
-package qna.domain;
+package qna.domain.answer;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import qna.domain.user.User;
 
 import java.util.List;
 import java.util.Optional;
