+++ src/main/java/qna/domain/question/QuestionRepository.java
@@ -1,4 +1,4 @@
-package qna.domain;
+package qna.domain.question;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
