+++ src/main/java/qna/domain/user/UserRepository.java
@@ -1,4 +1,4 @@
-package qna.domain;
+package qna.domain.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
